fix: Report all errors
At each phase of processing, all errors are returned.  For `stage`, the
application bails out at the end of that processing phase.

My hope was to keep the `ErrorPredicate` and `Errors` decoupled.  My
hope had been it would build up the errors internally and that a user
could then access it after.  The problem is that `collect` takes
ownership of the iterator.

Fixes #3

BREAKING CHANGE: `builder::Staging` changed from a type def to a
distinct type.